requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
pillow==12.0.0
sqlparse==0.5.5
tzdata==2025.3
